// src/store/enums/userSetting.enum.ts
export enum UserSetting {
	CB_USERNAME = 'cbUsername',
	CB_PASSWORD = 'CB_PASSWORD',
	CB_ADDRESS = 'cbAddress',
	ADVANCED_SEARCH_ENABLED = 'ADVANCED_SEARCH_ENABLED',
	SELECTED_TRACKER = 'SELECTED_TRACKER',
	CBQL_STRING = 'CBQL_STRING',
	SUBQUERY_LINK_METHOD = 'SUBQUERY_LINK_METHOD',
	FILTER_CRITERIA = 'FILTER_CRITERIA',
}